Rearrange some files.
Add new windows test module for doing testsuites.
