+++ test/new-e2e/tests/sysprobe-functional/assets/iisstart.png
